requirements.txt
PyQt6==6.7.0
PyQt6-WebEngine==6.7.0
markdown2==2.4.13
python-slugify==8.0.4
sqlalchemy==2.0.36